refactor: move concepts to their own files
sendLogstream and link_poll_event have been moved to their respective
files
